fix auth email otp, recode, sort project tree
