+++ .env
@@ -29,7 +29,7 @@ REFRESH_TOKEN_SECRET = 'myrefreshtoken'
 HOST_SMTP = 'smtp.gmail.com'
 GOOGLE_MAILER_CLIENT_ID = '251967640945-nnrjjfil8gjer0cpp7d1fd4rlghnhhs4.apps.googleusercontent.com'
 GOOGLE_MAILER_CLIENT_SECRET = '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'
-GOOGLE_MAILER_REFRESH_TOKEN = '1//04R7U-HdncDtaCgYIARAAGAQSNwF-L9Irtm9ovw34Kz0kR_8XU9ApIU2BOAVN9elSYThGAECK8O4-uzZjPLOYrlocYJy7Rbxtw7A'
+GOOGLE_MAILER_REFRESH_TOKEN = '1//04HLws5uBFidmCgYIARAAGAQSNwF-L9IrEYTbSk8rL0jpTzmjmzhCEKilDVMccbABiHe7gaK5uNEkECfkDKx5caItQWW8Yo31Wk0'
 ADMIN_EMAIL_ADDRESS = 'phamngocson7a1@gmail.com'
 ADMIN_PASS_ADDRESS = 'pvtxekqmvnuxdtvk'
