+++ server.js
@@ -10,12 +10,11 @@ const path = require("path");
 const cookieParser = require("cookie-parser");
 const passport = require("passport");
 const cors = require("cors");
-const userRoute = require("./route/users");
-const authRoute = require("./route/auth");
-const postRoute = require("./route/post");
-const notifyRoute = require("./route/notify");
-const chatRoute = require("./route/chat");
-
+const userRoute = require("./routes/users");
+const authRoute = require("./routes/auth");
+const postRoute = require("./routes/post");
+const notifyRoute = require("./routes/notify");
+const chatRoute = require("./routes/chat");
 
 // ENV Config
 dotenv.config();
@@ -23,11 +22,9 @@ const port = process.env.PORT || 8080;
 
 //MongoDB Connect
 mongoose.connect(process.env.MONGO_URL, (err) => {
-    if (err) console.log(err);
-    else console.log("Connected to MongoDB");
+  if (err) console.log(err);
+  else console.log("Connected to MongoDB");
 });
-
-
 
 //MIDDLEWARE
 // app.use(
@@ -35,35 +32,35 @@ mongoose.connect(process.env.MONGO_URL, (err) => {
 // );
 // app.use(session({ secret: 'sonlody' }));
 
-app.use(express.json())
-app.use(helmet())
-app.use(morgan("dev"))
-app.use(bodyParser.json())
-app.use(cookieParser())
+app.use(express.json());
+app.use(helmet());
+app.use(morgan("dev"));
+app.use(bodyParser.json());
+app.use(cookieParser());
 app.use(passport.initialize());
 
-
-
-app.use(cors({
-  origin: true,
-  credentials: true
-}))
+app.use(
+  cors({
+    origin: true,
+    credentials: true,
+  })
+);
 
 //ROUTE
 app.use("/api/auth", authRoute);
 app.use("/api/user", userRoute);
 app.use("/api/post", postRoute);
-app.use("/api/notify", notifyRoute);
-app.use("/api/chat", chatRoute);
+// app.use("/api/notify", notifyRoute);
+// app.use("/api/chat", chatRoute);
 
-app.use(express.static(path.join(__dirname, 'build')));
-app.get('/*', (req, res) => {
-  res.sendFile(path.join(__dirname, 'build', 'index.html'));
+app.use(express.static(path.join(__dirname, "build")));
+app.get("/*", (req, res) => {
+  res.sendFile(path.join(__dirname, "build", "index.html"));
 });
 // Serve static assets if in production
 // if (process.env.NODE_ENV === 'production') {
 //   // Set static folder
-  // app.use(express.static('client/build'));
+// app.use(express.static('client/build'));
 
 //   app.get('*', (req, res) => {
 //     res.sendFile(path.resolve(__dirname, 'client', 'build', 'index.html'));
@@ -71,7 +68,6 @@ app.get('/*', (req, res) => {
 // }
 
 app.listen(port, () => console.log("Server is running on port is:", port));
-
 
 // COOKIES SESSION
 
@@ -81,4 +77,3 @@ app.listen(port, () => console.log("Server is running on port is:", port));
 //   // Cookie Options
 //   maxAge: 24 * 60 * 60 * 1000 // 24 hours
 // }))
-
